Исправление замечаний 26.11.2024
