+++ src/main/java/org/example/Main.java
@@ -1,61 +1,48 @@
 package org.example;
 
-import org.example.lessonFour.exerciseEight.ExerciseEight;
-import org.example.lessonFour.exerciseEleven.ExerciseEleven;
-import org.example.lessonFour.exerciseFive.ExerciseFive;
-import org.example.lessonFour.exerciseFour.ExerciseFour;
-import org.example.lessonFour.exerciseFourteen.ExerciseFourteen;
-import org.example.lessonFour.exerciseNine.ExerciseNine;
-import org.example.lessonFour.exerciseOne.ExerciseOne;
-import org.example.lessonFour.exerciseSeven.ExerciseSeven;
-import org.example.lessonFour.exerciseSix.ExerciseSix;
-import org.example.lessonFour.exerciseTen.ExerciseTen;
-import org.example.lessonFour.exerciseThirteen.ExerciseThirteen;
-import org.example.lessonFour.exerciseThree.ExerciseThree;
-import org.example.lessonFour.exerciseTwelve.ExerciseTwelve;
-import org.example.lessonFour.exerciseTwo.ExerciseTwo;
+import org.example.lessonFour.MethodsFofLessonFour;
 
 import java.util.Scanner;
 
-//TIP To <b>Run</b> code, press <shortcut actionId="Run"/> or
-// click the <icon src="AllIcons.Actions.Execute"/> icon in the gutter.
 public class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-         //Задание 1.
-        new ExerciseOne().printThreeWords();
+        MethodsFofLessonFour method = new MethodsFofLessonFour();
+
+        //Задание 1.
+        method.printThreeWords();
         // Задание 2.
-        new ExerciseTwo().checkSumSign();
+        method.checkSumSign();
         // Задание 3.
-        new ExerciseThree().printColor();
+        method.printColor();
         // Задание 4.
-        new ExerciseFour().compareNumbers();
+        method.compareNumbers();
         //Задание 5.
-        new ExerciseFive().checkNumbers();
-         //Задание 6.
-        System.out.println("Задание 6.\n Введите целое число: ");
+        method.checkNumbers();
+        //Задание 6.
+        System.out.println("Задание 6.\nВведите целое число: ");
         int a = scanner.nextInt();
-        new ExerciseSix().positiveOrNegative(a);
+        method.positiveOrNegative(a);
         // Задание 7.
-        System.out.println("Задание 7.\n  Введите целое число: ");
+        System.out.println("Задание 7.\nВведите целое число: ");
         int b = scanner.nextInt();
-        new ExerciseSeven().returnResult(b);
+        method.returnResult(b);
         // Задание 8.
-        new ExerciseEight().printLine("Привет всем!", 5);
+        method.printLine("Привет всем!", 5);
         //Задание 9.
-        System.out.println("Задание 9.\n  Введите год: ");
+        System.out.println("Задание 9.\nВведите год: ");
         int year = scanner.nextInt();
-        new ExerciseNine().leapYear(year);
+        method.leapYear(year);
         //Задание 10.
-        new ExerciseTen().arrayChange();
+        method.arrayChange();
         //Задание 11.
-        new ExerciseEleven().arrayLengthHundred();
+        method.arrayLengthHundred();
         //Задание 12.
-        new ExerciseTwelve().correctArray();
+        method.correctArray();
         //Задание 13.
-        new ExerciseThirteen().findElementsArray();
+        method.findElementsArray();
         //Задание 14.
-        new ExerciseFourteen().lenInitialValue(5, 100);
+        method.lenInitialValue(5, 100);
         scanner.close();
     }
 }
