--- src/main/java/org/example/lessonFour/exerciseFour/ExerciseFour.java
@@ -1,15 +1,0 @@
-package org.example.lessonFour.exerciseFour;
-/*Метод, который сравнивает два числа, и выводит сообщение "a >= b" или "a < b" */
-public class ExerciseFour {
-    public void compareNumbers() {
-        int a = 3;
-        int b = 4;
-
-        if (a >= b) {
-            System.out.println("Задание 4: a >= b");
-        }
-        else {
-            System.out.println("Задание 4: a < b");
-        }
-    }
-}
