--- src/main/java/org/example/lessonFour/exerciseOne/ExerciseOne.java
@@ -1,14 +1,0 @@
-package org.example.lessonFour.exerciseOne;
-
-/*Метод который при вызове отпечатывает в втолбец три слова:
-* Orange, Banana, Apple. */
-
-public class ExerciseOne {
-
-    public void printThreeWords() {
-        System.out.println("Задание 1:");
-        System.out.println("Orange");
-        System.out.println("Banana");
-        System.out.println("Apple");
-    }
-}
