--- src/main/java/org/example/lessonFour/exerciseSix/ExerciseSix.java
@@ -1,14 +1,0 @@
-package org.example.lessonFour.exerciseSix;
-
-import java.util.Scanner;
-/*Метод, проверящий, явяляется ли число введеное с консоли, положительным или отрицательным. */
-public class ExerciseSix {
-    public void positiveOrNegative(int a) {
-        if (a >= 0) {
-            System.out.println("Ответ: Число положительно");
-        }
-        else {
-            System.out.println("Ответ: Число отрицательное");
-        }
-    }
-}
